requirements.txt
fastapi>=0.115.0
uvicorn>=0.32.0
jinja2>=3.1.4
python-multipart>=0.0.9
pydantic>=2.9.0
requests>=2.32.0
beautifulsoup4>=4.12.3
python-dotenv>=1.0.1
openai>=1.54.0
crawl4ai>=0.4.0
qdrant-client>=1.11.0
